# GoPalette/api/comment/v1/comment_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: comment/v1/comment.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationCommentCreateComment = "/api.comment.v1.Comment/CreateComment"
const OperationCommentDeleteComment = "/api.comment.v1.Comment/DeleteComment"
const OperationCommentListComments = "/api.comment.v1.Comment/ListComments"

type CommentHTTPServer interface {
	// CreateComment 发表评论/回复
	CreateComment(context.Context, *CreateCommentRequest) (*CommentInfo, error)
	// DeleteComment 删除评论
	DeleteComment(context.Context, *DeleteCommentRequest) (*DeleteCommentReply, error)
	// ListComments 拉取文章评论列表（按一级评论分页，附带二级回复）
	ListComments(context.Context, *ListCommentsRequest) (*ListCommentsReply, error)
}

func RegisterCommentHTTPServer(s *http.Server, srv CommentHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/comments", _Comment_CreateComment0_HTTP_Handler(srv))
	r.GET("/v1/posts/{post_id}/comments", _Comment_ListComments0_HTTP_Handler(srv))
	r.DELETE("/v1/comments/{id}", _Comment_DeleteComment0_HTTP_Handler(srv))
}

func _Comment_CreateComment0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateCommentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentCreateComment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateComment(ctx, req.(*CreateCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CommentInfo)
		return ctx.Result(200, reply)
	}
}

func _Comment_ListComments0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListCommentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentListComments)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListComments(ctx, req.(*ListCommentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListCommentsReply)
		return ctx.Result(200, reply)
	}
}

func _Comment_DeleteComment0_HTTP_Handler(srv CommentHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteCommentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCommentDeleteComment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteComment(ctx, req.(*DeleteCommentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteCommentReply)
		return ctx.Result(200, reply)
	}
}

type CommentHTTPClient interface {
	// CreateComment 发表评论/回复
	CreateComment(ctx context.Context, req *CreateCommentRequest, opts ...http.CallOption) (rsp *CommentInfo, err error)
	// DeleteComment 删除评论
	DeleteComment(ctx context.Context, req *DeleteCommentRequest, opts ...http.CallOption) (rsp *DeleteCommentReply, err error)
	// ListComments 拉取文章评论列表（按一级评论分页，附带二级回复）
	ListComments(ctx context.Context, req *ListCommentsRequest, opts ...http.CallOption) (rsp *ListCommentsReply, err error)
}

type CommentHTTPClientImpl struct {
	cc *http.Client
}

func NewCommentHTTPClient(client *http.Client) CommentHTTPClient {
	return &CommentHTTPClientImpl{client}
}

// CreateComment 发表评论/回复
func (c *CommentHTTPClientImpl) CreateComment(ctx context.Context, in *CreateCommentRequest, opts ...http.CallOption) (*CommentInfo, error) {
	var out CommentInfo
	pattern := "/v1/comments"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCommentCreateComment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteComment 删除评论
func (c *CommentHTTPClientImpl) DeleteComment(ctx context.Context, in *DeleteCommentRequest, opts ...http.CallOption) (*DeleteCommentReply, error) {
	var out DeleteCommentReply
	pattern := "/v1/comments/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentDeleteComment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListComments 拉取文章评论列表（按一级评论分页，附带二级回复）
func (c *CommentHTTPClientImpl) ListComments(ctx context.Context, in *ListCommentsRequest, opts ...http.CallOption) (*ListCommentsReply, error) {
	var out ListCommentsReply
	pattern := "/v1/posts/{post_id}/comments"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCommentListComments))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}