# GoPalette/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/comments:
        post:
            tags:
                - Comment
            description: 发表评论/回复
            operationId: Comment_CreateComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.comment.v1.CreateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.comment.v1.CommentInfo'
    /v1/comments/{id}:
        delete:
            tags:
                - Comment
            description: 删除评论
            operationId: Comment_DeleteComment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.comment.v1.DeleteCommentReply'
...
    /v1/posts/{postId}/comments:
        get:
            tags:
                - Comment
            description: 拉取文章评论列表（按一级评论分页，附带二级回复）
            operationId: Comment_ListComments
            parameters:
                - name: postId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.comment.v1.ListCommentsReply'
...
components:
    schemas:
        api.comment.v1.AuthorInfo:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                avatarUrl:
                    type: string
        api.comment.v1.CommentInfo:
            type: object
            properties:
                id:
                    type: string
                postId:
                    type: string
                userId:
                    type: string
                content:
                    type: string
                parentId:
                    type: string
                rootId:
                    type: string
                likeCount:
                    type: string
                status:
                    type: integer
                    format: enum
                author:
                    $ref: '#/components/schemas/api.comment.v1.AuthorInfo'
                replyToAuthor:
                    $ref: '#/components/schemas/api.comment.v1.AuthorInfo'
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                replies:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.comment.v1.CommentInfo'
        api.comment.v1.CreateCommentRequest:
            type: object
            properties:
                postId:
                    type: string
                content:
                    type: string
                parentId:
                    type: string
        api.comment.v1.DeleteCommentReply:
            type: object
            properties:
                success:
                    type: boolean
        api.comment.v1.ListCommentsReply:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.comment.v1.CommentInfo'
                total:
                    type: string
...